Rework SAFT to operate in verified rootfs environment.
With the advent of verified rootfs SAFT became unusable, as
it is remounting the root file system in rw mode to install
an upstart script.

The solution is to have an immutable upstart script (as
being done under ttp://codereview.chromium.org/3330005/show)
and to modify the SAFT framework to support that, which is
being done by this CL.

The immutable script tries to find the shell script
installed during SAFT initialization. The shell script is
installed on the flash drive's stateful partition. in
/var/saft.sh.

To make sure that the file names used by the immutable
upstart script and SAFT stay in sync, the upstart script is
 defining the vital file names through <name>=<value>
statements, and SAFT parses the upstart script to retrieve
 these values.

A few bugs were uncovered while testing these changes, they
are being fixed too.

Since the names of the

BUG=chromium-os:6345
TEST=manual

The tests pass, the actual test log will be added after
finalizing the upstart script changes.

Review URL: http://codereview.chromium.org/3297002

+++ saft/chromeos_interface.py
@@ -92,7 +92,8 @@ class ChromeOSInterface(object):
 
     def shut_down(self, new_log='/var/saft_log.txt'):
         '''Destroy temporary environment so that the test can be restarted.'''
-        shutil.copyfile(self.log_file, new_log)
+        if os.path.exists(self.log_file):
+            shutil.copyfile(self.log_file, new_log)
         shutil.rmtree(self.state_dir)
 
     def log(self, text):
@@ -241,7 +242,7 @@ class ChromeOSInterface(object):
             dev = '/dev/root'
 
         for line in self.run_shell_command_get_output('mount'):
-            if not line.startswith(dev):
+            if not line.startswith('%s ' % dev):
                 continue
             mount_options = options_filter.match(line).groups(0)[0]
         # found mounted
